refactor: collocate interface declarations
